refactor(ipa): random_with_eval inline
