+++ benches/pcs.rs
@@ -8,6 +8,7 @@ use arecibo::traits::{
   TranscriptEngineTrait,
 };
 use criterion::{criterion_group, criterion_main, Bencher, BenchmarkId, Criterion, SamplingMode};
+use ff::Field;
 use halo2curves::bn256::Bn256;
 use rand::rngs::StdRng;
 use rand_core::{CryptoRng, RngCore, SeedableRng};
@@ -49,12 +50,33 @@ struct BenchAssests<E: Engine, EE: EvaluationEngineTrait<E>> {
   proof: Option<<EE as EvaluationEngineTrait<E>>::EvaluationArgument>,
 }
 
+/// Returns a random polynomial, a point and calculate its evaluation.
+pub fn random_poly_with_eval<E: Engine, R: RngCore + CryptoRng>(
+  num_vars: usize,
+  mut rng: &mut R,
+) -> (
+  MultilinearPolynomial<<E as Engine>::Scalar>,
+  Vec<<E as Engine>::Scalar>,
+  <E as Engine>::Scalar,
+) {
+  // Generate random polynomial and point.
+  let poly = MultilinearPolynomial::random(num_vars, &mut rng);
+  let point = (0..num_vars)
+    .map(|_| <E as Engine>::Scalar::random(&mut rng))
+    .collect::<Vec<_>>();
+
+  // Calculation evaluation of point over polynomial.
+  let eval = MultilinearPolynomial::evaluate_with(poly.evaluations(), &point);
+
+  (poly, point, eval)
+}
+
 impl<E: Engine, EE: EvaluationEngineTrait<E>> BenchAssests<E, EE> {
-  pub(crate) fn from_num_vars<R: CryptoRng + RngCore>(ell: usize, mut rng: &mut R) -> Self {
-    let (poly, point, eval) = MultilinearPolynomial::random_with_eval(ell, &mut rng);
+  pub(crate) fn from_num_vars<R: CryptoRng + RngCore>(num_vars: usize, mut rng: &mut R) -> Self {
+    let (poly, point, eval) = random_poly_with_eval::<E, R>(num_vars, &mut rng);
 
     // Mock commitment key.
-    let ck = E::CE::setup(b"test", 1 << ell);
+    let ck = E::CE::setup(b"test", 1 << num_vars);
     // Commits to the provided vector using the provided generators.
     let commitment = E::CE::commit(&ck, poly.evaluations());
